+++ molo/surveys/views.py
@@ -13,16 +13,53 @@ from wagtail.wagtailcore.utils import cautious_slugify
 from django.contrib.auth.models import Group
 from django.core.urlresolvers import reverse
 from django.shortcuts import render
+from django.http import JsonResponse
 from django.utils.translation import ugettext as _
 
 from wagtail.wagtailadmin import messages
 from wagtail.wagtailadmin.utils import permission_required
+from wagtail_personalisation.forms import SegmentAdminForm
+from wagtail_personalisation.models import Segment
 
 from .forms import CSVGroupCreationForm
 
 
+class SegmentCountForm(SegmentAdminForm):
+    class Meta:
+        model = Segment
+        fields = ['type', 'status', 'count', 'name', 'match_any']
+
+
+def get_segment_user_count(request):
+        f = SegmentCountForm(request.POST)
+        context = {}
+        if f.is_valid():
+            rules = [
+                form.instance for formset in f.formsets.values()
+                for form in formset
+                if form not in formset.deleted_forms
+            ]
+            count = f.count_matching_users(rules, f.instance.match_any)
+            context = {'segmentusercount': count}
+        else:
+            errors = f.errors
+            # Get the errors for the Rules forms
+            for formset in f.formsets.values():
+                if formset.has_changed():
+                    for form in formset:
+                        if form.errors:
+                            id_prefix = form.prefix
+                            for name, error in form.errors.items():
+                                input_name = id_prefix + "-%s" % name
+                                errors[input_name] = error
+
+            context = {'errors': errors}
+
+        return JsonResponse(context)
+
+
 class SurveySuccess(TemplateView):
-    template_name = "surveys/molo_survey_page_landing.html"
+    template_name = "surveys/molo_survey_page_success.html"
 
     def get_context_data(self, *args, **kwargs):
         context = super(TemplateView, self).get_context_data(*args, **kwargs)
@@ -67,7 +104,7 @@ class SurveySuccess(TemplateView):
             for question, answers in results.items():
                 total = sum(answers.values())
                 for key in answers.keys():
-                    answers[key] = (answers[key] * 100) / total
+                    answers[key] = int((answers[key] * 100) / total)
         context.update({'self': survey, 'results': results})
         return context